refactor: Port→domain/repository、Adapter→infrastructure/persistenceへ配置変更
