+++ backend-go/internal/domain/repository/game_repository.go
@@ -1,4 +1,4 @@
-package ports
+package repository
 
 import (
 	"context"
